refactor:축제리스트의 컴포넌트화,api get 요청 test 상품 안보이는 곳까지 불러오기
