+++ moamoa/src/Recoil/ProductTypeStateAtom.jsx
@@ -1,0 +1,11 @@
+import { atom } from 'recoil';
+
+export const festivalActiveState = atom({
+  key: 'festivalActiveState',
+  default: true,
+});
+
+export const experienceActiveState = atom({
+  key: 'experienceActiveState',
+  default: false,
+});
